toolkit/mozapps/downloads/tests/chrome/test_esc_key_closes_clears.xul

<?xml version="1.0"?>
<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/.  -->
<!--
 * This test ensures that the escape key closes the download manager.  This test
 * was added in bug 416303.
-->

<window title="Download Manager Test"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="test();">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>
  <script type="application/javascript"
          src="utils.js"/>

  <script type="application/javascript">
  <![CDATA[

const DLMGR_UI_DONE = "download-manager-ui-done";

let os = Cc["@mozilla.org/observer-service;1"].
         getService(Ci.nsIObserverService);

function test()
{
  var dmui = getDMUI();
  if (!dmui) {
    todo(false, "skip test for toolkit download manager UI");
    return;
  }

  // Close the UI if necessary
  let wm = Cc["@mozilla.org/appshell/window-mediator;1"].
           getService(Ci.nsIWindowMediator);
  let win = wm.getMostRecentWindow("Download:Manager");
  if (win) win.close();

  // Register with the observer service
  os.addObserver(testObs, DLMGR_UI_DONE, false);

  // Show the Download Manager UI
  dmui.show();

  SimpleTest.waitForExplicitFinish();
}

let testObs = {
  observe: function(aSubject, aTopic, aData)
  {
    if (aTopic != DLMGR_UI_DONE)
      return;

    SimpleTest.waitForFocus(function () { continueTest(aSubject) }, aSubject);
  }
};

function continueTest(win)
{
  // Put in a value to clear out
  let search = win.document.getElementById("searchbox");
  search.focus();
  search.value = "download manager escape test";

  // Send the escape key to the download manager UI window
  let sendEscape = function() synthesizeKey("VK_ESCAPE", {}, win);

  // Escape from search box
  sendEscape();
  is(win.closed, false,
    "Escape doesn't close the window when in the search box");
  is(search.value, "", "Escape correctly emptied the search box");

  // Escape from list (escape from search box moves focus to list)
  sendEscape();
  is(win.closed, true,
    "Previous escape moved focus to list and now escape closed the window");

  os.removeObserver(testObs, DLMGR_UI_DONE);
  SimpleTest.finish();
};

  ]]>
  </script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display:none;"></div>
    <pre id="test"></pre>
  </body>
</window>
